requirements.txt
aiogram
pyrogram
pydantic
pydantic-settings
python-dotenv
sqlalchemy
aiosqlite
apscheduler
black[d]
geopy
haversine